+++ src/main/java/com/wenxin2/marioverse/datagen/EntityTypeTagsGen.java
@@ -536,7 +536,8 @@ public class EntityTypeTagsGen extends EntityTypeTagsProvider {
         tag(TagRegistry.CHECKPOINT_FLAG_CANNOT_SPAWN);
 
         tag(TagRegistry.CHEEP_CHEEPS)
-                .add(EntityRegistry.CHEEP_CHEEP.get());
+                .add(EntityRegistry.CHEEP_CHEEP.get())
+                .add(EntityRegistry.SPINY_CHEEP_CHEEP.get());
 
         tag(TagRegistry.CHEEP_CHEEP_CAN_ATTACK)
                 .add(EntityType.PLAYER)
@@ -841,6 +842,14 @@ public class EntityTypeTagsGen extends EntityTypeTagsProvider {
         tag(TagRegistry.SPIKE_PANEL_IMMUNE)
                 .addTag(TagRegistry.POWER_UP_ENTITIES)
                 .add(EntityRegistry.BOO.get());
+
+        tag(TagRegistry.SPINY_CHEEP_CHEEP_CAN_ATTACK)
+                .add(EntityType.PLAYER)
+                .add(EntityType.VILLAGER)
+                .add(EntityType.WANDERING_TRADER)
+                .addOptional(GUARD_VILLAGER)
+                .addOptional(FEMALE_VILLAGER)
+                .addOptional(MALE_VILLAGER);
 
         tag(TagRegistry.SPLUNKIN_CAN_ATTACK)
                 .add(EntityType.PLAYER);